#2 Lay groundwork for current labelling
Terminals have been upgraded to 'first-class citizens', in that they now are VMobjects in their own right and are administered seperately from a component's body.

This is to allow current annotations to be attached to terminals rather than components. This approach lends particular flexibility for the future introduction of components with more than two poles.