#2 Add submobject processing overrides
The `__rotate` and `__marks` groups are now private and hidden as an implementation detail of the Markable class, forcing interaction with its functionality from subclasses to be cleaner.

The Markable class's `add()`, `add_to_back()` , and `remove()` methods have been overriden so as to correctly handle the `__mark` and `__rotate` subgroups of it (so that if a Markable is added as a submobject of another Markable the marks still don't rotate), and otherwise just add the mobject to the `__rotate` group.