#2 Fix tests
The refactors of the previous commits broke some tests and made others impossible (as they relied on implementation details of the Component class that had been moved). These have been fixed, and tests added for Mark and Markable to test their functionality (some of which was tested in test_component.py when that functionality was in Component).