#2 Add current clear animations
Also added a consideration for terminal arrangement when placing the arrow, so that it will now properly respect the `out` parameter (in that it will always point the right way).

Also also explicitly told the Uncreate animation in the override for `_clear_mark()` not to handle removal, as the override handles it itself.