Reorganise utils
Moved the utils out of the _utils module's __init__.py and into its own utils.py file. This makes them easier to find and the file easier to identify in a tab, as well as allowing the utils to be imported as 'utils' whilst still using the preferred relative imports, maintaining a closer link to their origin.

Also removed the utils function `small_values_zeroed_out()` as it was unused outside of the testing framework.