#7 Reorganise directories
The 'components' package was increasingly containing things that weren't components (terminals, marks, markables). It was decided to move these to the '_base' package, and to leave 'components' to be just for actual, usable, instantiatable components. This is in preparation for the introduction of a Voltage class, which had to be put somewhere sensible.

The standard import method has also been changed to be fully qualified from the package root (i.e. starting with `manim_eng` and not relying on any reimports), to make it more obvious where a given unit is being imported from.