#7 Add updater for voltage
This now means that the attached terminals and/or the avoided components can move and otherwise change and the arrow will respond accordingly.

There is currently a visual glitch where the old form of the arrow is still shown until the end of the animation, which it would be good to get rid of.