#7 Introduce voltage constructing method and move Voltage to public module.
The design for the voltage setting on components was changed, as having components being able to manage their own voltage arrows was leading to a great deal of complexity, both in the implementation and in the userside API. It now just returns a newly-made voltage object, and expects the user to do with that what they wish.

The Voltage class has been moved to a public module 'circuit', as it is designed to be used by an end user. This name was chosen as this will be where the still-to-be-created circuit mechanism will go.

The Poetry lockfile was also updated.