#8 Clean up disconnection methods
Some of the common behaviour of the disconnection methods has been abstracted to make maintenance easier (in particular, maintenance of the animation overrides to keep them in line with the main methods).

The signature of `.disconnect()` has additionally been changed to mirror that of `.isolate()`, i.e. to allow whole components to be passed as well as terminals.