Rename test files to have 'test' at the end
This was done to make them less confusing, e.g. 'test_utils' was a test set for manim_eng's utils, but looks like a set of test utilities, and 'test_component' was a test set for the Component class, but looks like a test component. The rename alleviates this confusion.