#9 Change Terminal interface
The 'position' argument to the Terminal constructor now refers to the point the terminal attaches to the component it belongs to, *not* the point other things connect to it at. This was done because it's much more intuitive, and all the places calling the constructor were having to do extra calculations to work out where the end of the terminal should be. Specifying the start instead keeps everything much cleaner and easier to understand.