#9 Move structure around a little
Base and helper classes for components have been moved to the `base` subpackage of `components`, so that they are together with the actual components. It was also decided to expose these to the external interface, so that the user can define their own components as they wish.

The Bipole class was moved out of component.py and Markable out of mark.py to keep the files shorter and more specific, as well as to make the files containing these classes easier to locate.