#9 Add current and voltage sources
This uses a new, almost trait-like architecture for encapsulating common aspects of components. It relies on inheritance-based running of all construction methods. This was done to make it very easy to define lots of different sources that all have common characteristics to some of the others (European/American, controlled/uncontrolled).

The Poetry lockfile was also updated.