#10 Rename voltage methods
These are to make them easier to find. A new method was also introduced as the obvious counterpart to the renamed `.set_clockwise()`.

The (now) `.set_label()` method also now accepts a `clockwise` argument, as it is likely both may need to be set (and *animated*) at once. The animation override has been updated to include this.