#18 Add Node component
Also refactored Component to keep terminals in their own VGroup, so that Node can more easily handle adding terminals to/from the terminal list. A property ``.terminals`` was added to Component to maintain the old interface of this attribute being a list of the component's terminals (as this is still a useful internal and user-facing API).