#18 Adjust node type setting behaviour
This is now more aligned to a logical set of default behaviours.

The `.set_open()` method has been completely removed, as it was of limited use, and was having to take many arguments that were only relevant to one case of the two. In other words, it was screaming to be two separate dedicated methods.