#27 Add to_latex() method on Unit, UnitSequence, Value
This is in preparation for the integration of `Value`s into `.set_label()` and `.set_annotation()`.

Also moved the units and utils test files back to the main tests directory as it made no sense for
(a) units to be in utils/ given that isn't where it is in the source anymore; or
(b) utils to be in a directory on its own.