Fix debug mode leakage
The environment used by the manim directive is leaky and shared between invocations. This can lead to the debug mode leaking into other visualisations.

Though it's not pretty, the best solution I've found so far is to explicitly set the debug mode to False at the start of each manim directive invocation which doesn't require the debug mode.